Student Random Quiz

+++ app/src/main/java/uk/ac/aber/dcs/cs31620/quizapp/fragments/teacher/add/questions/AddQuestions.kt
@@ -63,7 +63,7 @@ class AddQuestions : Fragment() {
                 args.currentModuleName,
                 args.currentQuestionName,
                 questionName,
-                questionAnswer,
+                questionAnswer.toInt(),
                 optionAns1,
                 optionAns2,
                 optionAns3,
